fix: Address code review issues from PR #116
- Fix failing property-based test by filtering reserved device names
- Extract duplicate is_output_pin() function to shared utils module
- Optimize has_device_to_device check by combining loops in validation

All tests now pass (794 passed). Changes improve code quality by:
1. Preventing test data generation conflicts with reserved names
2. Following DRY principle by eliminating duplicate helper functions
3. Improving performance by reducing redundant iterations
